Function zng_ext_single_instance::single_instance

pub fn single_instance()

Enable single instance mode for the app-process.

The current executable path is used as the unique name. See single_instance_named for more details about name.

Examples

The example below demonstrates a single instance process setup. The single_instance() function is called before the app starts building. After the app starts building, before run, a subscription for APP_INSTANCE_EVENT is setup, this event will receive args for the current instance on run and for other instances latter.

fn main() {
    view_process::init();
    // zng::task::ipc::run_worker(worker);

    // must be called after `view_process`, `run_worker` and before the APP build.
    zng::app::single_instance();

    app_main();
}

fn app_main() {
    let app = APP.defaults();

    zng::app::APP_INSTANCE_EVENT
        .on_event(app_hn!(|args: &zng::app::AppInstanceArgs, _| {
            println!("app instance #{}, args: {:?}", args.count, args.args);
        }))
        .perm();

    app.run_window(async {
        Window! {
            child_align = Align::CENTER;
            child = Button! {
                child = Text!("Spawn Instance");
                on_click = hn!(|_| {
                    let exe = std::env::current_exe().unwrap();
                    std::process::Command::new(exe).arg("app arg 1").arg("--arg2").spawn().unwrap();
                });
            }
        }
    });
}