Function zng_ext_single_instance::single_instance_named

pub fn single_instance_named(unique_name: impl Into<Txt>)

Enable single instance mode for the app-process, using a custom unique identifier for the app.

The name should have only ASCII alphanumerics and be at maximum 128 bytes in length only, otherwise it will be mangled. The name is used to create a global lock, see the single-instance crate for more details.